+++ lib/parse/model/core/properties.rb
@@ -277,6 +277,11 @@ module Parse
           # Also autofetch if object was partially fetched and this field wasn't included.
           should_autofetch = value.nil? && (pointer? || (partially_fetched? && !field_was_fetched?(key)))
           if should_autofetch
+            # If autofetch is disabled and we're accessing an unfetched field on a
+            # partially fetched object, raise an error to make the issue explicit
+            if autofetch_disabled? && partially_fetched? && !field_was_fetched?(key)
+              raise Parse::UnfetchedFieldAccessError.new(key, self.class.name)
+            end
             # call autofetch to fetch the entire record
             # and then get the ivar again cause it might have been updated.
             autofetch!(key)